% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ex6crossing.R
\docType{data}
\name{Ex6crossing}
\alias{Ex6crossing}
\title{Time-to-event data example 6 for non-proportional hazards working group}
\format{Data frame with 4 variables:
\describe{ 
\item{id}{sequential numbering of unique identifiers}
\item{month}{time-to-event}
\item{event}{1 for event, 0 for censored} 
\item{trt}{1 for experimental, 0 for control}
}}
\usage{
data(Ex6crossing)
}
\description{
Survival objects reverse-engineered datasets from published Kaplan-Meier
curves. 
Individual trials are de-identified since the data are only
approximations of the actual data.
Data are intended to evaluate methods and designs for trials where
non-proportional hazards may be anticipated for outcome data.
}
\examples{
data(Ex6crossing)
library(survival)
km1 <- with(Ex6crossing,survfit(Surv(month,evntd)~trt))
km1
plot(km1)
with(subset(Ex6crossing,trt==1),pwexpfit(Srv=Surv(month,evntd),3))
with(subset(Ex6crossing,trt==0),pwexpfit(Srv=Surv(month,evntd),3))
}
\references{
TBD
}
\seealso{
\code{\link{Ex1delayedEffect}}, \code{\link{Ex2delayedEffect}}, \code{\link{Ex3belly}},
\code{\link{Ex4belly}}, \code{\link{Ex5widening}}
}
\keyword{datasets}
