% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsUpdate.R
\name{gsUpdate}
\alias{gsUpdate}
\title{Update gsDesign object boundary}
\usage{
gsUpdate(dth, d, b)
}
\arguments{
\item{dth}{Vector with number of events}

\item{d}{object of class 'gsDesign'}

\item{b}{1 for upper bound, 0 for lower bound}
}
\description{
Update gsDesign object boundary based on actual number of events. 
Returns Z-values for upper or lower bounds for an updated design.
Internal function only.
}
\details{
Details to be added
}
\keyword{internal}
