% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmk.Stat.r
\name{lmk.Stat}
\alias{lmk.Stat}
\title{Landmark analysis}
\usage{
lmk.Stat(survival, cnsr, trt, stra = NULL, fparam)
}
\arguments{
\item{survival}{time-to-event variable}

\item{cnsr}{censoring variable: 1=censoring, 0=event}

\item{trt}{treatment varaible. Accepted values are either "experiment" or "control"}

\item{stra}{stratification variable. Default is \code{NULL} (currently not implemented)}

\item{fparam}{a list input
\describe{
 \item{\code{fparam$lmk}}{time at which the landmark analysis is done} 
 \item{\code{fparam$lmktype}}{test name of which the p-Value will be stored in the variable \code{pval}. 
 Accepted values are "naive", "log", "loglog", "arcsin", "logit"}
 }}
}
\value{
The function return a list with the follow components
\describe{
 \item{pval}{One-sided p-Value from user specified test} 
 \item{pval_[XXX]}{One-sided test from various transformations}
 }
}
\description{
Landmark (survival curve comparison at fixed point in time) analysis based on Klein 2007 (statistics in Medicine)
}
\details{
The p-Values from the following transformations of survival are included:
\describe{
 \item{naive}{No transformation} 
 \item{log}{\code{log(S)}}
 \item{loglog}{\code{log(-log(S))}}
 \item{arcsin}{\code{arcsin(sqrt(S))}}
 \item{logit}{\code{log(S/(1-S))}}
 }
}
\examples{
# Lankmark analysis on the simulated data
library(survMisc)
medC = 6 
hr <- c(1, 0.6)
intervals <- 3 
gamma <- c(2.5, 5,  7.5,  10) ## a ramp-up enrollment
R     <- c(2  , 2,  2  ,  6 ) ## enrollment period: total of 12 months
eta <- -log(0.99) ## 1\% monthly dropout rate
sim1 <- nphsim(nsim=1,lambdaC=log(2)/medC,lambdaE=log(2)/medC*hr, ssC=300,ssE=300,intervals=intervals,gamma=gamma, R=R,eta=eta)
test1 <- simtest(x=sim1, anaD=c(250,300), method=lmk.Stat,fparam=list(lmk=9, lmktype='loglog'))
test1$result[]

# direct function call (without cutoff)
lmk.Stat(surv=sim1$simd$survival, cnsr=sim1$simd$cnsr, trt=sim1$simd$treatment, fparam=list(lmk=9, lmktype='loglog'))

}
