% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nphAlltest.R
\name{nphAlltest}
\alias{nphAlltest}
\title{Title (for R Documentation)}
\usage{
nphAlltest(tte, eventflag, arm)
}
\arguments{
\item{tte}{time-to-event duration}

\item{eventflag}{status variable: 1=event, 0=censoring}

\item{arm}{treatment varaible. 1 = experimental arm, 0 = control arm}
}
\value{
The function return a table with the follow components
\describe{
 \item{column 1 (replace with actual column name)}{If you want to provide more details for each column in the return table, you can do it here.} 
 \item{column 2}{etc...}
 \item{column 3}{etc...}
 }
}
\description{
Description Text is put here.
}
\details{
Details text are provided here
}
\examples{
library(survival)
library(survMisc)
indf <- data.frame(trt=rep(0:1,100),aval=rexp(200,c(0.002,0.0018)),event=rbinom(200,1,0.9))
out <- nphAlltest(tte=indf$aval,eventflag=indf$event,arm=indf$trt)
out

}
