% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nSurvNPH.R
\name{nSurvNPH}
\alias{nSurvNPH}
\title{Sample Size Under Non-Propotional Hazards Assumptions}
\usage{
nSurvNPH(lambdaC = log(2)/6, hr = 0.6, hr0 = 1, etaC = 0, etaE = NULL,
  NONCMPL = 0, DROPIN = 0, gamma = 1, R = 12, S = NULL, T = NULL,
  minfup = NULL, ratio = 1, alpha = 0.025, beta = 0.1, sided = 1,
  tol = .Machine$double.eps^0.25, SBDV = 20, SIMULT = FALSE,
  TRT_LAG = 0, nstrata = 1)
}
\arguments{
\item{lambdaC}{scalar, vector or matrix of event hazard rates for the 
control group; rows represent time periods while columns represent strata; a vector implies a single stratum.}

\item{hr}{hazard ratio (experimental/control) under the alternate hypothesis (scalar)}

\item{hr0}{hazard ratio (experimental/control) under the null hypothesis (scalar)}

\item{etaC}{scalar, vector or matrix of dropout hazard rates for the control group; rows represent time periods while columns represent strata; if entered as a scalar, rate is constant accross strata and time periods; 
if entered as a vector, rates are constant accross strata}

\item{etaE}{matrix dropout hazard rates for the experimental group specified 
in like form as eta; if NULL, this is set equal to eta}

\item{NONCMPL}{TBD}

\item{DROPIN}{TBD}

\item{gamma}{a scalar, vector or matrix of rates of entry by time period (rows) 
and strata (columns); if entered as a scalar, rate is constant accross strata 
and time periods; if entered as a vector, rates are constant accross strata}

\item{R}{a scalar or vector of durations of time periods for recruitment rates 
specified in rows of gamma. Length is the same as number of rows in gamma. 
Note that when variable enrollment duration is specified (input T=NULL), 
the final enrollment period is extended as long as needed}

\item{S}{a scalar or vector of durations of piecewise constant event rates 
specified in rows of lambda, eta and etaE; this is NULL if there is a single 
event rate per stratum (exponential failure) or length of the number of rows 
in lambda minus 1, otherwise}

\item{T}{study duration; if T is input as NULL, this will be computed on output; 
see details.}

\item{minfup}{follow-up of last patient enrolled; if minfup is input as NULL, 
this will be computed on output; see details}

\item{ratio}{randomization ratio of experimental treatment divided by control; 
normally a scalar, but may be a vector with length equal to number of strata.}

\item{alpha}{type I error rate. Default is 0.025 since 1-sided testing is default}

\item{beta}{type II error rate. Default is 0.10 (90% power); 
NULL if power is to be computed based on other input values}

\item{sided}{1 for 1-sided testing, 2 for 2-sided testing}

\item{tol}{or cases when T or minfup values are derived through root finding (T or minfup input as NULL), 
tol provides the level of error input to the uniroot() root-finding function. The default is the same as for uniroot.}

\item{SBDV}{This is a tuning parameter for the Lakatos method specifying how many intervals to use for
the Markov approximation applied in the method}

\item{SIMULT}{TBD}

\item{TRT_LAG}{lag treatment}

\item{nstrata}{TBD}
}
\description{
Description text; note that parameter descriptions below as well as text for 
help file needs to be updated.
}
\details{
Details text
}
\examples{
#no staggered entry, no lag time;
#no staggered entry, lag time;
#stagger entry, no lag time
#stagged entry; lag time;
#strata;

}
