% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.sim.R
\name{plotsim}
\alias{plotsim}
\title{Plot Simulated Trial Outcome Summaries}
\usage{
plotsim(d, y, dg = NULL, yt = "Y Axis Title", logs = NULL, b = NULL,
  v = c(2, 0.5))
}
\arguments{
\item{d}{data.table "result" generated by simtest function (or a table table that has the same structure as \code{result})}

\item{y}{vector of the name(s) of the y varaible(s) for the boxplot. e.g. c("hr")}

\item{dg}{number of digits for median of y below the table. NULL will disable the display of the y statistics.}

\item{yt}{y-axis label}

\item{logs}{\code{TRUE}=log scale for y-axis}

\item{b}{whether to show boundary crossing probability at each analysis. This require xeff and xfut exist in the \code{result} data table}

\item{v}{height of graph area and the summary area below.}
}
\description{
Generate boxplot, display summary of number of death, timing, boundary crossing probability, and user requested simulation-level information
}
\details{
\code{plotsim()} takes the \code{result} data table from the return value of the \code{simtest()} function 
(or a table table that has the same structure as \code{result}) and generate the boxplot and display summary statistics.
}
\examples{
# Use a gsSurv object as both input for simulation and testing. A logrank p-value and HR from cox model is reported.
library(gsDesign)
gs <- gsSurv ( k = 3, test.type = 4, alpha = 0.025, beta = 0.05, timing = c( 0.5,0.75 ), sfu = sfHSD , sfupar = c( -4 ), sfl = sfHSD, sflpar = c( -12 ), lambdaC = log(2) / 6, hr = 0.65, hr0 = 1, eta = 0.01, gamma = c( 2.5,5,7.5,10 ), R = c( 2,2,2,6 ) , S = NULL , T = 15 , minfup = 3 , ratio = 1) 
sim1 <- nphsim(nsim=1000,d=gs)
test1 <- simtest(x=sim1,anatype='event',method='LR', d=gs)
test1$result
plotsim(test1$result,y=c("hr","pval"),dg=2,yt="Hazard Ratio",b=1,v=c(2,0.8)) ## not useful, demonstration purpose only 

}
