% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm.combo.WLRmax.R
\name{rm.combo.WLRmax}
\alias{rm.combo.WLRmax}
\title{Weighted Logrank combination with asymptotic p-value adjustment}
\usage{
rm.combo.WLRmax(time = NULL, status = NULL, arm = NULL, wt = NULL,
  adjust.methods = c("holm", "hochberg", "hommel", "bonferroni", "asymp")[3],
  ties.method = c("exact", "breslow", "efron")[2], one.sided = FALSE,
  HT.est = FALSE, max = TRUE, alpha = 0.025)
}
\arguments{
\item{time}{time-to-event variable}

\item{status}{event indicator: 0=censored, 1=event}

\item{arm}{treatment group indicator}

\item{wt}{a list with pairs of non-negative numbers setting rho and gamma for each 
Fleming-Harrington weighting sheme to be used}

\item{adjust.methods}{methods used to adjust Type I error for testing multiple weighted logrank tests}

\item{ties.method}{methods for dealing with tied event times}

\item{one.sided}{logical indicating 1-sided (TRUE) vs 2-sided (FALSE) testing}

\item{HT.est}{logical used when adjust.methods is "asymp", max is TRUE to create an estimate
of HR that is averaged across the specified tests}

\item{max}{indicator of extent of output (TRUE for maximum output)}

\item{alpha}{level for testing}
}
\value{
The function returns a list with the follow components
\describe{
 \item{pval}{One-sided p-Value corresponding to Zmax} 
 \item{rho, gamma}{rho and gamma corresponding to Zmax}
 \item{Zmax}{Zmax}
 }
}
\description{
Description
}
\details{
This routine provides several options for correcting p-values for the maximum of multiple 
weighted logrank test statistics.
The most favorable of these (i.e., minimally sufficient p-value correction) is provided by the "asymp" option which
uses a published asymptotic correlation between different weighted logrank tests (Karrison, 2016).
However, there are also more conservative (less powerful) options to correct using 
Holm, Hochberg, Hommel or Bonferroni corrections.


Note that the user needs to load the `dplyr` and `Matrix` packages since these have NOT been added to the imports list for the package.
The correlation between different weighted logrank test statistics noted in the Karrison (2016) is the same as the variance of a weighted
logrank statistic with the sum of rho values and sum of gamma values as the respective rho and gamma parameters.
}
\examples{

data(Ex1delayedEffect)
with(Ex1delayedEffect,rm.combo.WLRmax(time=month,status=evntd,arm=trt,adjust.methods="asymp",
                                   HT.est=TRUE,wt=list(a1=c(0,0),a2=c(0,1),a3=c(1,0),a4=c(1,1)))) 
with(Ex1delayedEffect,rm.combo.WLRmax(time=month,status=evntd,arm=trt,
                                   wt=list(a1=c(0,0),a2=c(0,1),a3=c(1,0),a4=c(1,1)))) 
                                   
}
\references{
TG Karrison (2016), Versatile tests for comparing survival curves based on weighted log-rank statistics, 
Stata Journal, 16:3 pp 678-690
}
\author{
Satrajit Roychoudhury, \email{satrajit.roychoudhury@pfizer.com}
}
