% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simtestp.R
\name{simtestp}
\alias{simtestp}
\title{Simulation of Time-to-Event P-Values}
\usage{
simtestp(x, anaT = NULL, anaD = NULL, anatype = "event",
  method = NULL, stratum = NULL, fparam = NULL, d = NULL)
}
\arguments{
\item{x}{list genearted by simulation}

\item{anaT}{Calendar time for reporting, testing, and boundary crossing probability when a gsDesign object is specified.}

\item{anaD}{Events driven for reporting, testing, and boundary crossing probability when a gsDesign object is specified.}

\item{anatype}{use specified timing(anaT) if 'calendar'; use specified # of events(anaD) if 'event'}

\item{method}{statistical testing on simulated data.}

\item{stratum}{TBD}

\item{fparam}{TBD}

\item{d}{A gsDesign object as input. Timing/Events and the boundary will be used.}
}
\value{
A
}
\description{
Description here
}
\details{
simulation test: provide # of events/timing at specified timing/# of events;
provide HR and p-value at specified timing/events;
provide boundary crossing probability 
(experimental: cannot update boundary at different # of events)
}
\examples{
# examples here


}
\keyword{Needs}
\keyword{internal}
\keyword{update!}
