% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.lr.r
\name{test.lr}
\alias{test.lr}
\title{LR Test Statistic}
\usage{
test.lr(surv, cnsr, trt, stra = NULL, fparam = NULL)
}
\arguments{
\item{cnsr}{censoring variable: 1=censoring, 0=event}

\item{trt}{treatment varaible. Accepted values are either "experiment" or "control"}

\item{stra}{stratification variable. Default is \code{NULL}}

\item{survival}{time-to-event variable}
}
\value{
The function return a list with the follow components
\describe{
 \item{pval}{One-sided p-Value from the logrank test} 
 \item{z}{z statistics from the logrank test}
 \item{hr}{Hazard ratio from the cox proportional hazard model}
 \item{sehr}{Standard error of the hazard ratio}
 }
}
\description{
Tests if there is a difference between two survival curves using logrank test. One sided p-value, Z statistics, HR from cox model and standard error
of the HR are generated.
}
\details{
test.lr is used when method='LR'
}
\examples{
# logrank test on the simulated data
library(survival)
medC = 6 
hr <- 0.7
intervals <- NULL 
gamma <- c(2.5, 5,  7.5,  10) ## a ramp-up enrollment
R     <- c(2  , 2,  2  ,  6 ) ## enrollment period: total of 12 months
eta <- -log(0.99) ## 1\% monthly dropout rate
sim1 <- nphsim(nsim=1,lambdaC=log(2)/medC,lambdaE=log(2)/medC*hr, ssC=300,ssE=300,
               intervals=intervals,gamma=gamma, R=R,eta=eta)
test1 <- simtest(x=sim1, anaD=300, method='LR')
test1$result[]

# direct function call (without cutoff)
test.lr(surv=sim1$simd$survival, cnsr=sim1$simd$cnsr, trt=sim1$simd$treatment)


}
