% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wlr.Stat.r
\name{wlr.Stat}
\alias{wlr.Stat}
\title{Weighted Logrank Test Statistic}
\usage{
wlr.Stat(survival, cnsr, trt, stra = NULL, fparam)
}
\arguments{
\item{survival}{time-to-event variable}

\item{cnsr}{censoring variable: 1=censoring, 0=event}

\item{trt}{treatment varaible. Accepted values are either "experiment" or "control"}

\item{stra}{stratification variable. Default is \code{NULL} (currently not implemented)}

\item{fparam}{a list input to request additional FH test and time of delayed separation in the simplified APPLE method
\describe{
 \item{\code{fparam$FH(rho, gamma)}}{request FH p-Value with rho and gamma (e.g. rho=0 and gamma=0 is logrank test)} 
 \item{\code{fparam$wlr}}{test name of which the p-Value will be stored in the variable \code{pval}. 
 Accepted values are "1", "n", "sqrtN", "S1", "S2", "FH01","FH10","FH11","FH","APPLE"}
 \item{\code{fparam$APPLE}}{time of delayed separation used in the simplified APPLE method}
 }}
}
\value{
The function return a list with the follow components
\describe{
 \item{pval}{One-sided p-Value from user specified test} 
 \item{pval_[XXX]}{One-sided test from various weighted tests}
 }
}
\description{
Weighted logrank adopted from survMisc package
}
\details{
The p-Values from the following weighted tests are included:
\describe{
 \item{1}{log-rank} 
 \item{n}{Gehan-Breslow generalized Wilcoxon}
 \item{sqrtN}{Tarone-Ware}
 \item{S1}{Peto-Peto's modified survival estimate}
 \item{S2}{modified Peto-Peto (by Andersen)}
 \item{FH01}{Fleming-Harrington(0,1)}
 \item{FH10}{Fleming-Harrington(1,0)}
 \item{FH11}{Fleming-Harrington(1,1)}
 \item{FH11}{Fleming-Harrington(1,1)}
 \item{FH}{Fleming-Harrington with user specified \code{rho} and \code{gamma}}
 \item{APPLE}{Simplified APPLE method with user specified late separation time}
 }
}
\examples{
# weighted logrank on the simulated data
library(survMisc)
medC = 6 
hr <- c(1, 0.6)
intervals <- 3 
gamma <- c(2.5, 5,  7.5,  10) ## a ramp-up enrollment
R     <- c(2  , 2,  2  ,  6 ) ## enrollment period: total of 12 months
eta <- -log(0.99) ## 1\% monthly dropout rate
sim1 <- nphsim(nsim=1,lambdaC=log(2)/medC,lambdaE=log(2)/medC*hr, ssC=300,ssE=300,intervals=intervals,gamma=gamma, R=R,eta=eta)
test1 <- simtest(x=sim1, anaD=c(250,300), method=wlr.Stat,fparam=list(FH=c(0,2), wlr='1', APPLE=3))
test1$result[]

# direct function call (without cutoff)
wlr.Stat(surv=sim1$simd$survival, cnsr=sim1$simd$cnsr, trt=sim1$simd$treatment,fparam=list(FH=c(0,2), wlr='1', APPLE=3))

}
