% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nphAlltest.R
\name{nphAlltest}
\alias{nphAlltest}
\title{nphAlltest A Wrapper Function}
\usage{
nphAlltest(survival, cnsr, trt, rgs = list(c(0, 0), c(0, 1), c(0, 2), c(0, 3),
  c(1, 0), c(2, 0), c(2, 2)))
}
\arguments{
\item{survival}{time-to-event variable}

\item{cnsr}{censoring variable: 1=censoring, 0=event}

\item{trt}{treatment variable}
}
\value{
The function returns a list with the following components
\describe{
 \item{lr}{raw output from log-Rank test} 
 \item{cox}{raw output from cox PH model}
 \item{maxWLR}{raw output from max-combo test by Larry}
 \item{comboLR}{raw output from max-WLR test by Satrajit}
 \item{rmst}{raw output from restricted mean test}
 \item{summary.stat}{concatenated summary statistics including 1-sided p-value, point estimate (mean difference/ratio for RMST tests and hazard ratios for the rest tests), lower and upper bound of 95 percent confidence interval}
 }
}
\description{
A wrapper function to run all candidate tests on given data
}
\details{
The function runs the following tests then consolidates key statistics into a summary table. \cr
Log-Rank test/cox proportional hazard model, max-combo test by Larry, Max-WLR Test by Satrajit and restricted mean test.
}
\examples{
library(survival)
library(survMisc)
library(Matrix)
library(dplyr)
set.seed(1)
indf <- data.frame(trt=rep(c('control','experimental'),100),aval=rexp(200,c(0.002,0.0018)),
                   event=rbinom(200,1,0.9))
out <- nphAlltest(survival=indf$aval,cnsr=1-indf$event,trt=indf$trt)
out$summary.stat
out <- nphAlltest(survival=indf$aval,cnsr=1-indf$event,trt=indf$trt,rgs=list(c(0,0),c(0,1)))
out

}
