% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wlr.Stat.r
\name{wlr.Stat}
\alias{wlr.Stat}
\title{Weighted Logrank Test Statistic}
\usage{
wlr.Stat(survival, cnsr, trt, stra = NULL, fparam)
}
\arguments{
\item{survival}{time-to-event variable}

\item{cnsr}{censoring variable: 1=censoring, 0=event}

\item{trt}{treatment varaible. Accepted values are either "experiment" or "control"}

\item{stra}{stratification variable. Default is \code{NULL} (currently not implemented)}

\item{fparam}{a list input to request additional FH test and time of delayed separation in the simplified APPLE method
\describe{
 \item{\code{fparam$rho}}{a vector of \code{rho} to be used in Fleming-Harrington FH(rho, gamma)} 
 \item{\code{fparam$gamma}}{a vector of \code{gamma} to be used in Fleming-Harrington \code{FH(rho, gamma)}. 
 The length of \code{gamma} needs to be the same as \code{rho}} 
 \item{\code{fparam$wlr}}{test name of which the p-Value will be stored in the variable \code{pval}. 
 Accepted values are "1", "n", "sqrtN", "S1", "S2", "APPLE", or "FH(rho,gamma)" where \code{rho} and \code{gamma} is one of the
 pair specified by user in the \code{fparam}}
 \item{\code{fparam$APPLE}}{time of delayed separation used in the simplified APPLE method}
 }}
}
\value{
The function return a list with the follow components
\describe{
 \item{pval}{One-sided p-Value from user specified test} 
 \item{pval_[XXX]}{One-sided test from various weighted tests}
 }
}
\description{
Weighted logrank adopted from survMisc package
}
\details{
The p-Values from the following weighted tests are included:
\describe{
 \item{1}{log-rank} 
 \item{n}{Gehan-Breslow generalized Wilcoxon}
 \item{sqrtN}{Tarone-Ware}
 \item{S1}{Peto-Peto's modified survival estimate}
 \item{S2}{modified Peto-Peto (by Andersen)}
 \item{APPLE}{Simplified APPLE method with user specified late separation time}
 \item{FH(rho,gamma)}{Fleming-Harrington with user specified pairs of \code{rho} and \code{gamma}}
 }
}
\examples{
# weighted logrank on the simulated data
library(survMisc)
medC = 6 
hr <- c(1, 0.6)
intervals <- 3 
gamma <- c(2.5, 5,  7.5,  10) ## a ramp-up enrollment
R     <- c(2  , 2,  2  ,  6 ) ## enrollment period: total of 12 months
eta <- -log(0.99) ## 1\% monthly dropout rate
sim1 <- nphsim(nsim=2,lambdaC=log(2)/medC,lambdaE=log(2)/medC*hr, ssC=300,ssE=300,intervals=intervals,gamma=gamma, R=R,eta=eta)
test1 <- simtest(x=sim1, anaD=c(250,300), method=wlr.Stat,fparam=list(rho=c(0,0,1), gamma=c(0,1,1), wlr='FH(0,1)', APPLE=3))
test1$result[]

# direct function call (without cutoff)
wlr.Stat(surv=sim1$simd$survival, cnsr=sim1$simd$cnsr, trt=sim1$simd$treatment,fparam=list(rho=c(0,0,1), gamma=c(0,1,1), wlr='FH(0,1)', APPLE=3))

}
